requirements.txt
Django==1.8.5
SPARQLWrapper==1.7.6
coverage==4.0.2
django-jenkins==0.18.1
docutils==0.12
html5lib==0.9999999
isodate==0.5.4
keepalive==0.5
pyparsing==2.0.7
python-dateutil==2.4.2
rdflib==4.2.1
simplejson==3.8.2
six==1.10.0
wsgiref==0.1.2
